Refactor del refreshToken y aplicados a todas las interacciones que requieran un usuario logeado

+++ slim/src/Modules/UsersModule.php
@@ -114,7 +114,7 @@ final class UsersModule {
     }
 
     //si el token es valido, refresco su expiración
-    \Authentication::refreshToken($db, $id, 300);
+    \Authentication::refreshToken($db, $auth['id'], 300);
 
     //Escribe en el body la respuesta en formato JSON
     $res->getBody()->write(json_encode($row));
@@ -288,7 +288,7 @@ final class UsersModule {
               ->fetch(\PDO::FETCH_ASSOC);
 
     //si el token es valido, refresco su expiración
-    \Authentication::refreshToken($db, $id, 300);
+    \Authentication::refreshToken($db, $auth['id'], 300);
 
     $res->getBody()->write(json_encode($row, JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES));
     return $res->withHeader('Content-Type','application/json; charset=utf-8');
@@ -299,6 +299,9 @@ final class UsersModule {
   // DELETE /user/{id} -------------------------------------------------------------------------------
   public static function deleteUser(Request $req, Response $res, array $args): Response {
     $targetId = (int)($args['id']);//id desde la ruta
+    
+    
+
 
     //Valido si la id es menor o igual a 0
     if ($targetId <= 0) {
@@ -318,7 +321,7 @@ final class UsersModule {
     $db = \DB::getConnection();
 
     //si el token es valido, refresco su expiración
-    \Authentication::refreshToken($db, $id, 300);
+    \Authentication::refreshToken($db, $auth['id'], 300);
 
     //busco el usuario a borrar
     $row = $db->query("SELECT id, is_admin FROM users WHERE id = $targetId LIMIT 1")
@@ -338,14 +341,14 @@ final class UsersModule {
                     ->withStatus(400);
     }
 
-    //ACA VERIFICARIA SI EL USUARIO TIENE RESERVAS
-    //SELECT 1 FROM bookings WHERE user_id = $targetId LIMIT 1
-    //$hasBookings = $db->query("SELECT 1 FROM bookings WHERE user_id = $targetId LIMIT 1")->fetchColumn();
-    //if ($hasBookings) {
-    //  $res->getBody()->write(json_encode(['error' => 'El usuario tiene reservas y no se puede eliminar']));
-    //  return $res->withHeader('Content-Type','application/json; charset=utf-8')->withStatus(409);
-    //}
-    //------------------------------------------------------------
+    //Verifico si el usuario tiene reservas
+    $hasBookings = $db->query("SELECT 1 FROM bookings WHERE created_by = $targetId LIMIT 1")
+                  ->fetchColumn();
+
+    if ($hasBookings) {
+      $res->getBody()->write(json_encode(['error' => 'El usuario posee reservas, no se puede eliminar']));
+      return $res->withHeader('Content-Type','application/json; charset=utf-8')->withStatus(409);
+    }
 
     //si no tiene reservas, lo borro
     $db->exec("DELETE FROM users WHERE id = $targetId LIMIT 1");
